requirements.txt
jupedsim>=1.3.2
pedpy>=1.2.0
shapely>=2.0.0
numpy>=2.2.0
pandas>=2.2.0
pyyaml>=6.0
scipy>=1.13.0
matplotlib>=3.8.0
